Fixed call vote button & review comments
