+++ Resources/Locale/en-US/_FarHorizons/voting/vote-manager.ftl
@@ -1,4 +1,5 @@
-﻿ui-vote-faction-title = Next faction
+﻿ui-vote-type-faction = Next faction
+ui-vote-faction-title = Next faction
 ui-vote-faction-tie = Tie when voting for the faction! A random faction is selected...
 ui-vote-faction-win = Voting for the faction is over!
 ui-vote-faction-notlobby = Voting for faction is only valid in the pre-round lobby!
